botones gestion pacientes, dashboard pacientes,
mejorada vista para admin y user en pacientes, editar arreglado (falta impactar en base de datos)

+++ views/paciente/gestion.php
@@ -51,8 +51,14 @@
 					</div>
 				<?php endif; ?>
 				<?php Utils::deleteSession('delete'); ?>
-
-				<table class="table mt-4 table-bordered">
+				
+				<?php if (isset($_SESSION['update']) && $_SESSION['update'] == 'complete') : ?>
+					<div class="alert alert-info my-4" role="alert">
+						<strong>El paciente ha sido actualizado correctamente</strong>
+					</div>
+				<?php endif; ?>
+				<?php Utils::deleteSession('update'); ?>
+				<table class="table table-striped table-hover">
 					<thead class="thead-light">
 						<tr>
 							<th>Ver</th>
@@ -66,14 +72,14 @@
 						</tr>
 					</thead>
 					<tbody>
-						<?php  // instanciamos obrasocial
+						<?php
 							require_once 'models/obrasociales.php';
 							$obrasocial = new ObraSocial();
 							?>
-					<?php while ($pac = $pacientes->fetch_object()) : ?>
-						<?php if (Utils::isAdmin()) : ?> <!-- comprobamos si es 'admin' y listamos todas las entradas de la tabla pacientes  -->
+						<?php while ($pac = $pacientes->fetch_object()) : ?>
+						<?php if ($pac->habilitado || Utils::showAdmin()) :?>
 							<tr>
-								<th>O</th>
+								<th><a href="<?=base_url?>paciente/dashboard&id=<?=$pac->id_paciente?>"><button type="button" class="btn btn-warning">dashboard</button></a></th>
 								<td scope="row"><?= $pac->id_paciente; ?></td>
 								<td><?= $pac->apellido . ", " . $pac->nombre; ?></td>
 								<td><?= $pac->dni; ?></td>
@@ -88,47 +94,25 @@
 
 								<td><?= $pac->sexo; ?></td>
 
-								<td><?php
-											$obra = $obrasocial->getObraSocial($pac->id_paciente);
-											echo $obra->nombre . " | Numero: ";
-											?>
+								<td>
+									<?php
+										$obra = $obrasocial->getObraSocial($pac->id_paciente);
+										echo $obra->nombre . " | Numero: ";
+									?>
 								</td>
 								<td>
-
-									<a href="<?= base_url ?>paciente/eliminar&id=<?= $pac->id_paciente ?>"><button type="button" class="btn btn-danger">Eliminar</button></a>
-
+									<?php if ($pac->habilitado) :?>
+										<a href="<?=base_url?>paciente/editar&id=<?=$pac->id_paciente?>"><button type="button" class="btn btn-warning">Editar</button></a>
+										<a href="<?= base_url ?>paciente/eliminar&id=<?= $pac->id_paciente ?>"><button type="button" class="btn btn-danger">Eliminar</button></a>
+									<?php elseif (!$pac->habilitado || Utils::showAdmin()) : ?>
+										<a href="<?= base_url ?>paciente/habilitar&id=<?= $pac->id_paciente ?>"><button type="button" class="btn btn-secondary">Habilitar</button></a>
+									<?php endif; ?>
 								</td>
 							</tr>
-							<?php elseif (!Utils::isAdmin() && $pac->habilitado) : ?> <!-- comprobamos que *no* sea admin y listamos solo los pacientes habilitados. Flag 1 en la base de datos -->
-							<tr>
-								<th>O</th>
-								<td scope="row"><?= $pac->id_paciente; ?></td>
-								<td><?= $pac->apellido . ", " . $pac->nombre; ?></td>
-								<td><?= $pac->dni; ?></td>
-								<td>
-									<?php
-										$cumpleanos = new DateTime($pac->fecha_nacimiento);
-										$hoy = new DateTime();
-										$annos = $hoy->diff($cumpleanos);
-										echo $annos->y;;
-									?>
-								</td>
-								<td><?= $pac->sexo; ?></td>
-								<td><?php
-											$obra = $obrasocial->getObraSocial($pac->id_paciente);
-											echo $obra->nombre . " | Numero: ";
-											?>
-								</td>
-								<td>
-									<a href="<?= base_url ?>paciente/eliminar&id=<?= $pac->id_paciente ?>"><button type="button" class="btn btn-danger">Eliminar</button></a>
-								</td>
-							</tr>	
-							<?php endif; ?> <!-- termina el if de la vista pacientes -->
-							<?php endwhile ; ?>
-
-
-						</body> </table> </div> </div> </div> </div> </div> <?php else : ?> <div class="col-10">
-
+							<?php endif; ?>
+						<?php endwhile; ?>
+						<t/body> </table> </div> </div> </div> </div> </div> <?php else : ?> <div class="col-10">
+						
 							<div class="alert alert-warning m-5" role="alert">
 								<strong>Necesitas ser andministrador.</strong> Inicia sesion aqui <a href="<?= base_url ?>"></a>
 							</div>
